[cpu_backend] Split Q4_0 GEMV kernels by architecture
Move the canonical-row Q4_0/Q8_0 GEMV implementations out of the common quantization source and into the selected NEON, AVX2, fallback, and SVE backend sources.
